fix: float to int err

+++ src/plot_maker_alt.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 
-y = list(map(int, input("Masukkan nilai y dipisahkan oleh koma (contoh: 1,2,3): ").split(',')))
+y = list(map(float, input("Masukkan nilai y dipisahkan oleh koma (contoh: 1,2,3): ").split(',')))
 x = list(range(1, len(y) + 1))
 
 plt.plot(x, y, marker='o', linestyle='-', color='b')
